@@ -12,29 +12,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@angular/animations": "^19.0.5",
+    "@angular/cdk": "^19.0.4",
+    "@angular/common": "^19.0.5",
+    "@angular/compiler": "^19.0.5",
+    "@angular/core": "^19.0.5",
+    "@angular/forms": "^19.0.5",
+    "@angular/localize": "^19.0.5",
+    "@angular/platform-browser": "^19.0.5",
+    "@angular/platform-browser-dynamic": "^19.0.5",
+    "@angular/router": "^19.0.5",
+    "@ng-bootstrap/ng-bootstrap": "^18.0.0",
     "@ng-idle/core": "^13.0.0",
     "@popperjs/core": "^2.11.8",
     "@stomp/stompjs": "^7.0.0",
     "angular2-ladda": "^5.0.0",
     "bootstrap": "3.3.6",
-    "chart.js": "^3.4.0",
+    "chart.js": "^4.4.7",
     "font-awesome": "4.7.0",
     "ladda": "^2.0.3",
     "lodash.merge": "^4.6.2",
     "markdown-it": "^13.0.1",
     "mermaid": "~10.9.3",
     "monaco-editor": "^0.48.0",
-    "ng2-charts": "^3.1.2",
+    "ng2-charts": "^8.0.0",
     "pace-js": "1.2.4",
     "rxjs": "~7.8.0",
     "simple-datatables": "^8.0.0",
@@ -43,17 +44,17 @@
     "tinycon": "0.6.8",
     "tslib": "^2.8.1",
     "ws": "^8.17.0",
-    "zone.js": "~0.14.4"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.11",
-    "@angular-eslint/builder": "^17.5.2",
-    "@angular-eslint/eslint-plugin": "^17.5.2",
-    "@angular-eslint/eslint-plugin-template": "^17.5.2",
-    "@angular-eslint/schematics": "^17.5.2",
-    "@angular-eslint/template-parser": "^17.3.12",
-    "@angular/cli": "^17.3.11",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular-devkit/build-angular": "^19.0.6",
+    "@angular-eslint/builder": "^19.0.2",
+    "@angular-eslint/eslint-plugin": "^19.0.2",
+    "@angular-eslint/eslint-plugin-template": "^19.0.2",
+    "@angular-eslint/schematics": "^19.0.2",
+    "@angular-eslint/template-parser": "^19.0.2",
+    "@angular/cli": "^19.0.6",
+    "@angular/compiler-cli": "^19.0.5",
     "@types/jasmine": "~4.6.4",
     "@types/lodash": "^4.17.14",
     "@types/markdown-it": "^13.0.7",
@@ -73,6 +74,6 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
     "prettier": "^3.4.2",
-    "typescript": "5.4.x"
+    "typescript": "5.6.3"
   }
 }
