@@ -2,9 +2,10 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { PagesNavigationComponent } from './pages-navigation.component';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { provideRouter } from '@angular/router';
-import { routes } from '../../../app-routing.module';
+import { routes } from '../../../app.routes';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('PagesNavigationComponent', () => {
   let component: PagesNavigationComponent;
@@ -12,9 +13,9 @@ describe('PagesNavigationComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [PagesNavigationComponent, HttpClientTestingModule],
-      providers: [provideRouter(routes)],
       schemas: [NO_ERRORS_SCHEMA],
+      imports: [PagesNavigationComponent],
+      providers: [provideRouter(routes), provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()],
     }).compileComponents();
 
     fixture = TestBed.createComponent(PagesNavigationComponent);
