@@ -2,13 +2,12 @@ import { AfterViewInit, Component, EventEmitter, Input, OnChanges, OnInit, Outpu
 import { convertToParamMap, Router, RouterModule } from '@angular/router';
 import { MinimalizaSidebarComponent } from './minimaliza-sidebar.component';
 import { ScrollToTopComponent } from './scroll-to-top.component';
-import { CommonModule } from '@angular/common';
 import { CustomViewsComponent } from '../../custom-views/custom-views.component';
-import { InformationModalComponent } from '../information-modal/information-modal.component';
 import { ServerInfoService } from '../../../services/server-info.service';
 import { CdkAccordionItem, CdkAccordionModule } from '@angular/cdk/accordion';
 import { SidebarDirective } from './sidebar.directive';
 import { HasAccessToLinkDirective } from '../../has-access-to-link.directive';
+import { NgClass } from '@angular/common';
 
 type ExpandedItem = {
   element: HTMLElement;
@@ -19,17 +18,15 @@ type ExpandedItem = {
   selector: 'app-pages-navigation',
   templateUrl: './pages-navigation.component.html',
   styleUrls: ['./pages-navigation.component.scss'],
-  standalone: true,
   imports: [
-    CommonModule,
     RouterModule,
     CdkAccordionModule,
     CustomViewsComponent,
     MinimalizaSidebarComponent,
     ScrollToTopComponent,
-    InformationModalComponent,
     SidebarDirective,
     HasAccessToLinkDirective,
+    NgClass,
   ],
 })
 export class PagesNavigationComponent implements OnChanges, OnInit, AfterViewInit {
