@@ -1,6 +1,8 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
-@Pipe({ name: 'variablesFilter' })
+@Pipe({
+  name: 'variablesFilter',
+})
 export class VariablesFilterPipe implements PipeTransform {
   transform<T>(variables: T[], filterText: string): T[] {
     const returnArray = [];
