@@ -1,8 +1,9 @@
 import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import type * as SockJS from 'sockjs-client';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import type SockJS from 'sockjs-client';
 
 import { WebsocketService } from './websocket.service';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 declare global {
   interface Window {
@@ -15,7 +16,8 @@ describe('WebsocketService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
+      imports: [],
+      providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()],
     });
     service = TestBed.inject(WebsocketService);
   });
