@@ -1,10 +1,11 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import { AppService } from 'src/app/app.service';
-import { BaseIframeComponent } from '../iframe.base';
+import { BaseIframeComponent, baseImports } from '../iframe.base';
 
 @Component({
   selector: 'app-iframe-ladybug',
+  imports: baseImports,
   templateUrl: '../iframe.component.html',
   styleUrls: ['../iframe.component.scss'],
 })
