@@ -1,14 +1,16 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { NotificationService, Notification } from 'src/app/services/notification.service';
+import { NgIf } from '@angular/common';
 
 @Component({
   selector: 'app-notifications',
+  imports: [NgIf],
   templateUrl: './notifications.component.html',
   styleUrls: ['./notifications.component.scss'],
 })
 export class NotificationsComponent implements OnInit {
-  protected notification: Notification | false = {
+  protected notification: Notification | null = {
     icon: '',
     title: '',
     message: false,
