@@ -1,9 +1,17 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { Router, RouterLink } from '@angular/router';
 import { AppService, JobMessage } from 'src/app/app.service';
 import { PollerService } from 'src/app/services/poller.service';
 import { SchedulerService, Trigger } from './scheduler.service';
 import { SweetalertService } from 'src/app/services/sweetalert.service';
+import { KeyValuePipe, NgClass, NgFor, NgIf } from '@angular/common';
+import { HasAccessToLinkDirective } from '../../components/has-access-to-link.directive';
+import { LaddaModule } from 'angular2-ladda';
+import { ToDateDirective } from '../../components/to-date.directive';
+import { TabListComponent } from '../../components/tab-list/tab-list.component';
+import { FormsModule } from '@angular/forms';
+import { SearchFilterPipe } from '../../pipes/search-filter.pipe';
+import { OrderByPipe } from '../../pipes/orderby.pipe';
 
 type Scheduler = {
   name: string;
@@ -36,6 +44,20 @@ export type Job = {
 
 @Component({
   selector: 'app-scheduler',
+  imports: [
+    NgIf,
+    HasAccessToLinkDirective,
+    LaddaModule,
+    ToDateDirective,
+    TabListComponent,
+    FormsModule,
+    KeyValuePipe,
+    SearchFilterPipe,
+    OrderByPipe,
+    NgFor,
+    RouterLink,
+    NgClass,
+  ],
   templateUrl: './scheduler.component.html',
   styleUrls: ['./scheduler.component.scss'],
 })
