@@ -1,12 +1,20 @@
 import { Component, OnDestroy, OnInit, TrackByFunction } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
-import { ConfigurationFilter } from 'src/app/pipes/configuration-filter.pipe';
+import { ConfigurationFilter, ConfigurationFilterPipe } from 'src/app/pipes/configuration-filter.pipe';
 import { StatusService } from './status.service';
 import { Adapter, AdapterStatus, Alert, AppService, Configuration, MessageLog } from 'src/app/app.service';
 import { PollerService } from 'src/app/services/poller.service';
 import { ServerInfo, ServerInfoService } from '../../services/server-info.service';
-import { KeyValue } from '@angular/common';
+import { KeyValue, KeyValuePipe, NgClass, NgFor, NgIf } from '@angular/common';
+import { ServerWarningsComponent } from './server-warnings/server-warnings.component';
+import { ConfigurationTabListComponent } from '../../components/tab-list/configuration-tab-list.component';
+import { ConfigurationSummaryComponent } from './configuration-summary/configuration-summary.component';
+import { HasAccessToLinkDirective } from '../../components/has-access-to-link.directive';
+import { FormsModule } from '@angular/forms';
+import { ConfigurationMessagesComponent } from './configuration-messages/configuration-messages.component';
+import { AdapterStatusComponent } from './adapter-status/adapter-status.component';
+import { SearchFilterPipe } from '../../pipes/search-filter.pipe';
 
 type Filter = Record<AdapterStatus, boolean>;
 
@@ -14,6 +22,21 @@ type Filter = Record<AdapterStatus, boolean>;
   selector: 'app-status',
   templateUrl: './status.component.html',
   styleUrls: ['./status.component.scss'],
+  imports: [
+    ServerWarningsComponent,
+    ConfigurationTabListComponent,
+    ConfigurationSummaryComponent,
+    HasAccessToLinkDirective,
+    NgIf,
+    NgClass,
+    FormsModule,
+    ConfigurationMessagesComponent,
+    AdapterStatusComponent,
+    NgFor,
+    ConfigurationFilterPipe,
+    SearchFilterPipe,
+    KeyValuePipe,
+  ],
 })
 export class StatusComponent implements OnInit, OnDestroy {
   protected filter: Filter = {
